Projeto adaptado as tabelas existentes no schema de operações

+++ src/main/java/br/edu/infnet/assessment/dto/RankingAventureiroDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,6 +13,6 @@ public class RankingAventureiroDTO {
     private Long idAventureiro;
     private String nomeAventureiro;
     private Long totalParticipacoes;
-    private Long somaRecompensas;
+    private BigDecimal somaRecompensas;
     private Long quantidadeDestaques;
 }
