+++ src/main/java/br/edu/infnet/assessment/dto/RelatorioMissaoDTO.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -14,5 +16,5 @@ public class RelatorioMissaoDTO {
     private StatusMissao status;
     private NivelPerigo nivelPerigo;
     private Long quantidadeParticipantes;
-    private Long totalRecompensasDistribuidas;
+    private BigDecimal totalRecompensasDistribuidas;
 }
